server/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
openai
pydantic
scikit-learn
pandas
numpy
datasets

